Fix some datatypes to help create persistent iSCSI disk
Fixed two datatypes to help create persistent iSCSI disk:
(1). remove the packageID 'omitempty' tag from the placeOrder datatype, as in iSCSI disk creation process, the packageId is 0 and SL need this packageId parameter in the placeOrder request.

(2). add datacenter attribute in the virtual guest datatype, as in the CPI, we need to get the datacenter id of the virtual guest to create an iSCSI disk.

These changes will not affect ephmeral disk feature, which we have tested.

Signed-off-by: Edward Zhang <zhuadl@cn.ibm.com>